add, delete 1 item product và URL traversal attack
Chỉnh lại delete item, và thêm tính năng add/delete 1 item cho product có sẵn. Và tạo thêm 1 function để check có phải admin không, nếu không thì không cho thực hiện lệnh. Ngoài ra còn chỉnh localhost qua 8082 vì 8081 để chạy trên docker rồi.

+++ demo/src/main/java/com/example/demo/controller/ProductController.java
@@ -1,7 +1,9 @@
 package com.example.demo.controller;
 
 import com.example.demo.entity.Product;
+import com.example.demo.entity.Users;
 import com.example.demo.service.ProductService;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,21 +22,50 @@ public class ProductController {
     }
 
     @GetMapping("/products")
-    public String listProducts(Model model) {
+    public String listProducts(@AuthenticationPrincipal Users user,Model model) {
+        if (checkAdmin(user)==false) { return "redirect:/cart"; }
         model.addAttribute("products", productService.getAllProducts());
         return "product-list"; // Trả về file product-list.html
     }
 
     @PostMapping("/products/add")
-    public String addProduct(@RequestParam String name, @RequestParam double price,
+    public String addProduct(@AuthenticationPrincipal Users user, @RequestParam String name, @RequestParam double price,
                              @RequestParam int quantity, @RequestParam String imageUrl) {
+        if (checkAdmin(user)==false) { return "redirect:/cart"; }
         productService.addProduct(new Product(name, price, quantity, imageUrl));
         System.out.println("Product added");
         return "redirect:/products"; // Chuyển hướng về trang danh sách
     }
-    @PostMapping("/delete/{id}")
-    public String deleteProduct(@PathVariable Long id) {
+    @PostMapping("/products/delete/{id}")
+    public String deleteProduct(@AuthenticationPrincipal Users user,@PathVariable Long id) {
+        if (checkAdmin(user)==false) { return "redirect:/cart"; }
         productService.deleteProductById(id);
         return "redirect:/products";
+    }
+
+    @PostMapping("/products/delete1item/{id}")
+    public String delete1ItemProduct(@AuthenticationPrincipal Users user,@PathVariable Long id) {
+        if (checkAdmin(user)==false) { return "redirect:/cart"; }
+        productService.delete1itemProductById(id);
+        return "redirect:/products";
+    }
+
+    @PostMapping("/products/add1item/{id}")
+    public String add1item(@AuthenticationPrincipal Users user,@PathVariable Long id) {
+        if (checkAdmin(user)==false) { return "redirect:/cart"; }
+        productService.add1item(id);
+        return "redirect:/products";
+    }
+
+    //@AuthenticationPrincipal
+    public boolean checkAdmin( Users user) {
+        String username = getContext().getAuthentication().getName();
+        if ("admin".equals(username)) {
+            return true;
+        }
+        else {
+            System.out.println("URL traversal attack");
+            return false;
+        }
     }
 }
